Move mvcore events handling to a seperate listener class
