+++ src/main/java/org/mvplugins/multiverse/inventories/handleshare/ShareHandleListener.java
@@ -75,61 +75,6 @@ public final class ShareHandleListener implements Listener {
         this.profileContainerStoreProvider = profileContainerStoreProvider;
     }
 
-    /**
-     * Adds Multiverse-Inventories version info to /mv version.
-     *
-     * @param event The MVVersionEvent that this plugin will listen for.
-     */
-    @EventHandler
-    void dumpsDebugInfoRequest(MVDumpsDebugInfoEvent event) {
-        event.appendDebugInfo(getDebugInfo());
-        File configFile = new File(this.inventories.getDataFolder(), "config.yml");
-        File groupsFile = new File(this.inventories.getDataFolder(), "groups.yml");
-        event.putDetailedDebugInfo("multiverse-inventories/config.yml", configFile);
-        event.putDetailedDebugInfo("multiverse-inventories/groups.yml", groupsFile);
-    }
-
-    /**
-     * Builds a String containing Multiverse-Inventories' version info.
-     *
-     * @return The version info.
-     */
-    private String getDebugInfo() {
-        StringBuilder versionInfo = new StringBuilder("[Multiverse-Inventories] Multiverse-Inventories Version: " + inventories.getDescription().getVersion() + '\n'
-                + "[Multiverse-Inventories] === Settings ===" + '\n'
-                + "[Multiverse-Inventories] First Run: " + config.getFirstRun() + '\n'
-                + "[Multiverse-Inventories] Using Bypass: " + config.getEnableBypassPermissions() + '\n'
-                + "[Multiverse-Inventories] Default Ungrouped Worlds: " + config.getDefaultUngroupedWorlds() + '\n'
-                + "[Multiverse-Inventories] Save and Load on Log In and Out: " + config.getSavePlayerdataOnQuit() + '\n'
-                + "[Multiverse-Inventories] Using GameMode Profiles: " + config.getEnableGamemodeShareHandling() + '\n'
-                + "[Multiverse-Inventories] === Shares ===" + '\n'
-                + "[Multiverse-Inventories] Optionals for Ungrouped Worlds: " + config.getUseOptionalsForUngroupedWorlds() + '\n'
-                + "[Multiverse-Inventories] Enabled Optionals: " + config.getActiveOptionalShares() + '\n'
-                + "[Multiverse-Inventories] === Groups ===" + '\n');
-
-        for (WorldGroup group : worldGroupManager.getGroups()) {
-            versionInfo.append("[Multiverse-Inventories] ").append(group.toString()).append('\n');
-        }
-
-        return versionInfo.toString();
-    }
-
-    @EventHandler
-    void onDebugModeChange(MVDebugModeEvent event) {
-        Logging.setDebugLevel(event.getLevel());
-    }
-
-    /**
-     * Hooks Multiverse-Inventories into the Multiverse reload command.
-     *
-     * @param event The MVConfigReloadEvent that this plugin will listen for.
-     */
-    @EventHandler
-    void configReload(MVConfigReloadEvent event) {
-        this.inventories.reloadConfig();
-        event.addConfig("Multiverse-Inventories - config.yml");
-    }
-
     @EventHandler(priority = EventPriority.MONITOR)
     void playerPreLogin(AsyncPlayerPreLoginEvent event) {
         if (event.getLoginResult() != Result.ALLOWED) {
@@ -339,6 +284,12 @@ public final class ShareHandleListener implements Listener {
         }
 
         World fromWorld = event.getFrom().getWorld();
+        if (fromWorld == null) {
+            // Apparently this happens sometimes.
+            Logging.fine("Entity %s attempted to go from null world", entity);
+            return;
+        }
+
         Location toLocation = event.getTo();
         if (toLocation == null) {
             // Apparently this happens sometimes.
@@ -359,8 +310,8 @@ public final class ShareHandleListener implements Listener {
         List<WorldGroup> fromGroups = worldGroupManager.getGroupsForWorld(fromWorld.getName());
         List<WorldGroup> toGroups = worldGroupManager.getGroupsForWorld(toWorld.getName());
         // We only care about the groups that have the inventory sharable
-        fromGroups = fromGroups.stream().filter(it -> it.isSharing(Sharables.INVENTORY)).collect(Collectors.toList());
-        toGroups = toGroups.stream().filter(it -> it.isSharing(Sharables.INVENTORY)).collect(Collectors.toList());
+        fromGroups = fromGroups.stream().filter(it -> it.isSharing(Sharables.INVENTORY)).toList();
+        toGroups = toGroups.stream().filter(it -> it.isSharing(Sharables.INVENTORY)).toList();
         for (WorldGroup fromGroup : fromGroups) {
             if (toGroups.contains(fromGroup)) {
                 Logging.finest("Allowing item or inventory holding %s to go from world %s to world %s", entity,
